+++ modules/paymill/metadata.php
@@ -12,6 +12,7 @@ $aModule = array(
     'email' => 'support@paymill.de',
     'extend' => array(
         'payment' => 'paymill/views/paymill_payment',
+        'order' => 'paymill/views/paymill_order',
         'oxpaymentgateway' => 'paymill/views/paymill_paymentgateway',
         'oxemail' => 'paymill/core/paymill_oxemail'
     ),
@@ -25,6 +26,7 @@ $aModule = array(
         'paymill_fastcheckout' => 'paymill/core/paymill_fastcheckout.php',
         'paymill_logging' => 'paymill/core/paymill_logging.php',
         'paymill_loglist' => 'paymill/core/paymill_loglist.php',
+        'paymill_logger' => 'paymill/core/paymill_logger.php',
         'paymill_util' => 'paymill/util/paymill_util.php',
         'paymill_hooks' => 'paymill/views/paymill_hooks.php'
     ),